mv btc and ltc unit test to work mod
